requirements.txt
async-timeout==4.0.3
bleak==0.22.3
joblib==1.5.1
numpy==2.2.6
paho-mqtt==2.1.0
pandas==2.3.1
pyModbusTCP==0.3.0
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-CoreBluetooth==10.3.2
pyobjc-framework-libdispatch==10.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
